Fixed the README shields
The build shield kept breaking due to being AppVeyor specific, this will allow me to recrete the build project as I am still playing around with it too much. The chocolatey shield has been added so that it can indicate the go live as soon as the package is approved.